BuildTool: Fix xcodebuild: error: SDK macosx26 cannot be located.
My SDKs directory contains three entries:

MacOSX.sdk
MacOSX26.0.sdk
MacOSX26.sdk

The current SDK version detection in hxcpp is preferring MacOSX26.sdk over MacOSX26.0.sdk. This is causing MACOSX_VER to be set to 26. However, xcodebuild seems to want both major and minor parts of the version to be specified, so MACOSX_VER should be set to 26.0 instead.

This change checks if the current best's minor version is parsed as NaN when the major versions match, which results in a real float value for the minor version to be preferred.